move functions from preprocessor to image_preprocessor and delete preprocessor
